+++ docs/js/main.js
@@ -8,17 +8,6 @@ import(componentUrl).then(res => {
 
   const $console = document.getElementById('console');
 
-  document.addEventListener('clipboard-copy-success', evt => {
-    console.log('clipboard-copy-success ->', evt.detail);
-    $console.innerHTML += `<div>$ <span class="success">clipboard-copy-success</span> -> ${JSON.stringify(evt.detail)}</div>`;
-
-    evt.target.querySelector('button').innerHTML = 'Copied!';
-
-    setTimeout(() => {
-      evt.target.querySelector('button').innerHTML = 'Copy';
-    }, 1000);
-  });
-
   document.addEventListener('clipboard-copy-error', evt => {
     console.log('clipboard-copy-error ->', evt.detail);
     $console.innerHTML += `<div>$ <span class="error">clipboard-copy-error</span> -> ${evt.detail.error.name}: ${evt.detail.error.message}</div>`;